+++ examples/surreal_ql/define_index.py
@@ -10,6 +10,8 @@ with Database("http://127.0.0.1:8000", 'test', 'test', credentials=('user_db', '
     print(db.define_index("userEmailIndex", "user").columns("email").unique().comment("unique index"))
     # DEFINE INDEX IF NOT EXISTS userEmailIndex ON TABLE user COLUMNS email UNIQUE;
     print(db.define_index("userEmailIndex", "user").if_not_exists().columns("email").unique())
+    # DEFINE INDEX OVERWRITE userEmailIndex ON TABLE user COLUMNS email UNIQUE;
+    print(db.define_index("userEmailIndex", "user").overwrite().columns("email").unique())
     # DEFINE INDEX userAgeIndex ON TABLE user COLUMNS age;
     print(db.define_index("userAgeIndex", "user").columns("age"))
     # DEFINE INDEX userNameIndex ON TABLE user COLUMNS name SEARCH ANALYZER ascii BM25 HIGHLIGHTS;