+++ src/surrealist/ql/statements/alter_statements.py
@@ -1,5 +1,0 @@
-from surrealist.ql.statements.statement import FinishedStatement, Statement
-
-
-
-